refactored the login modals
